Reduce steps to add blueprint to WeldWall
The new approach should be easier as you no longer have to make
sections in the custom data yourself. Another benefit is that now if
you change the projection offset before creating an entry, running
SAVE_OFFSET will actually work and there is no need to run PARSE which
resets the offset.